Add read method to GzipRecordFile and use it when parsing to improve parse time.
The read method is the same as in file objects. When reading in the
content of a warc record the parser now uses this method instead of
reading in a line oriented fashion. Reading lines required the use of a
regular expression which would be applied in an O(n^2) manner if the
current chunk did not contain a complete line. As a result reading
content which was not line oriented was extremely slow. The new method
does not involve parsing the content at all so is only limited by
decompression speed and Python loop overhead, to compensate for the
latter the CHUNK_SIZE has been increased.

CORE-465